mapauto-microservices/auto_nessus_service/requirements.txt
# =Juggernaut= Updated requirements.txt with all necessary dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.4.2
# Standard library modules (included in Python by default):
# os, json, typing
